samples/email_agent/requirements.txt
openai>=0.27.0
requests
google-auth>=2.20.0
google-auth-oauthlib>=0.8.0
google-api-python-client>=2.70.0
agent-framework